Rename React::API to React::ComponentFactory
