+++ lib/react/top_level.rb
@@ -1,5 +1,6 @@
 require "native"
 require 'active_support'
+require "react/ext/string"
 
 module React
   HTML_TAGS = %w(a abbr address area article aside audio b base bdi bdo big blockquote body br
@@ -23,7 +24,36 @@ module React
                 tabIndex target title type useMap value width wmode dangerouslySetInnerHTML)
 
   def self.create_element(type, properties = {}, &block)
-    React::API.create_element(type, properties, &block)
+    params = []
+
+    # Component Spec or Nomral DOM
+    if type.kind_of?(Class)
+      raise "Provided class should define `render` method"  if !(type.method_defined? :render)
+      params << React::ComponentFactory.native_component_class(type)
+    else
+      raise "#{type} not implemented" unless HTML_TAGS.include?(type)
+      params << type
+    end
+
+    # Passed in properties
+    props = {}
+    properties.map do |key, value|
+       if key == "class_name" && value.is_a?(Hash)
+         props[key.lower_camelize] = value.inject([]) {|ary, (k,v)| v ? ary.push(k) : ary}.join(" ")
+       else
+         props[React::ATTRIBUTES.include?(key.lower_camelize) ? key.lower_camelize : key] = value
+       end
+    end
+    params << props.shallow_to_n
+
+    # Children Nodes
+    if block_given?
+      children = [yield].flatten.each do |ele|
+        params << ele
+      end
+    end
+
+    return `React.createElement.apply(null, #{params})`
   end
 
   def self.render(element, container)
@@ -50,7 +80,7 @@ module React
 
   def self.expose_native_class(*args)
     args.each do |klass|
-      `window[#{klass.to_s}] = #{React::API.native_component_class(klass)}`
+      `window[#{klass.to_s}] = #{React::ComponentFactory.native_component_class(klass)}`
     end
   end
 end
